refactor: normalize require paths
