+++ src/reporters/x-unit-reporter.js
@@ -7,8 +7,8 @@ const file = require('file');
 const fsExtra = require('fs-extra');
 const pathmodule = require('path');
 
-const logger = require('./../logger');
-const prettifyResponse = require('./../prettify-response');
+const logger = require('../logger');
+const prettifyResponse = require('../prettify-response');
 
 function XUnitReporter(emitter, stats, tests, path, details) {
   EventEmitter.call(this);
