+++ src/transaction-runner.js
@@ -8,7 +8,7 @@ const { Pitboss } = require('pitboss-ng');
 
 const addHooks = require('./add-hooks');
 const logger = require('./logger');
-const packageData = require('./../package.json');
+const packageData = require('../package.json');
 const sortTransactions = require('./sort-transactions');
 const performRequest = require('./performRequest');
