+++ test/unit/reporters/html-reporter-test.js
@@ -10,7 +10,7 @@ const loggerStub = require('../../../src/logger');
 const fsExtraStub = { mkdirp(path, cb) { return cb(); } };
 
 const HtmlReporter = proxyquire('../../../src/reporters/html-reporter', {
-  './../logger': loggerStub,
+  '../logger': loggerStub,
   fs: fsStub,
   'fs-extra': fsExtraStub
 });
