+++ test/unit/reporters/x-unit-reporter-test.js
@@ -10,7 +10,7 @@ const loggerStub = require('../../../src/logger');
 const fsExtraStub = { mkdirp(path, cb) { return cb(); } };
 
 const XUnitReporter = proxyquire('../../../src/reporters/x-unit-reporter', {
-  './../logger': loggerStub,
+  '../logger': loggerStub,
   fs: fsStub,
   'fs-extra': fsExtraStub
 });
